Move changes to smpltmr
