+++ apps/smpltmr/ChangeLog
@@ -1,1 +1,2 @@
 0.01: Release
+0.02: Rewrite with new interface
