+++ apps/smpltmr/metadata.json
@@ -2,10 +2,10 @@
   "id": "smpltmr",
   "name": "Simple Timer",
   "shortName": "Simple Timer",
-  "version": "0.01",
+  "version": "0.02",
   "description": "A very simple app to start a timer.",
   "icon": "app.png",
-  "tags": "tool",
+  "tags": "tool,alarm,timer",
   "dependencies": {"scheduler":"type"},
   "supports": ["BANGLEJS2"],
   "screenshots": [{"url":"screenshot.png"}, {"url": "screenshot_2.png"}],
