--- apps/timersimple/ChangeLog
@@ -1,1 +1,0 @@
-0.01: First release
